requirements.txt
mysql-connector-python==8.0.33
pandas==2.0.3
python-dotenv==1.0.0
tqdm==4.66.1
